+++ rest_framework/fields.py
@@ -543,7 +543,13 @@ class HyperlinkedRelatedField(RelatedField):
         self.slug_url_kwarg = kwargs.pop('slug_url_kwarg', default_slug_kwarg)
 
         self.format = kwargs.pop('format', None)
+
+        self.use_absolute_urls = kwargs.pop('use_absolute_urls', None)
         super(HyperlinkedRelatedField, self).__init__(*args, **kwargs)
+
+    def initialize(self, parent, field_name):
+        super(HyperlinkedRelatedField, self).initialize(parent, field_name)
+        self.use_absolute_urls = self.use_absolute_urls or self.parent.use_absolute_urls
 
     def get_slug_field(self):
         """
@@ -555,12 +561,13 @@ class HyperlinkedRelatedField(RelatedField):
         view_name = self.view_name
         request = self.context.get('request', None)
         format = self.format or self.context.get('format', None)
+        use_absolute_urls = self.use_absolute_urls
         pk = getattr(obj, 'pk', None)
         if pk is None:
             return
         kwargs = {self.pk_url_kwarg: pk}
         try:
-            return reverse(view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass
 
@@ -571,13 +578,13 @@ class HyperlinkedRelatedField(RelatedField):
 
         kwargs = {self.slug_url_kwarg: slug}
         try:
-            return reverse(self.view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(self.view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass
 
         kwargs = {self.pk_url_kwarg: obj.pk, self.slug_url_kwarg: slug}
         try:
-            return reverse(self.view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(self.view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass
 
@@ -651,15 +658,21 @@ class HyperlinkedIdentityField(Field):
         self.pk_url_kwarg = kwargs.pop('pk_url_kwarg', self.pk_url_kwarg)
         self.slug_url_kwarg = kwargs.pop('slug_url_kwarg', default_slug_kwarg)
 
+        self.use_absolute_urls = kwargs.pop('use_absolute_urls', None)
         super(HyperlinkedIdentityField, self).__init__(*args, **kwargs)
+
+    def initialize(self, parent, field_name):
+        super(HyperlinkedIdentityField, self).initialize(parent, field_name)
+        self.use_absolute_urls = self.use_absolute_urls or self.parent.use_absolute_urls
 
     def field_to_native(self, obj, field_name):
         request = self.context.get('request', None)
         format = self.format or self.context.get('format', None)
         view_name = self.view_name or self.parent.opts.view_name
         kwargs = {self.pk_url_kwarg: obj.pk}
+        use_absolute_urls = self.use_absolute_urls
         try:
-            return reverse(view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass
 
@@ -670,13 +683,13 @@ class HyperlinkedIdentityField(Field):
 
         kwargs = {self.slug_url_kwarg: slug}
         try:
-            return reverse(self.view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(self.view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass
 
         kwargs = {self.pk_url_kwarg: obj.pk, self.slug_url_kwarg: slug}
         try:
-            return reverse(self.view_name, kwargs=kwargs, request=request, format=format)
+            return reverse(self.view_name, kwargs=kwargs, request=request, format=format, use_absolute_urls=use_absolute_urls)
         except:
             pass