+++ rest_framework/pagination.py
@@ -15,7 +15,11 @@ class NextPageField(serializers.Field):
             return None
         page = value.next_page_number()
         request = self.context.get('request')
-        url = request and request.build_absolute_uri() or ''
+        if self.parent.use_absolute_urls:
+            assert request, "request is required for building absolute url"
+            url = request.build_absolute_uri()
+        else:
+            url = request and request.get_full_path() or ''
         return replace_query_param(url, self.page_field, page)
 
 
@@ -30,7 +34,11 @@ class PreviousPageField(serializers.Field):
             return None
         page = value.previous_page_number()
         request = self.context.get('request')
-        url = request and request.build_absolute_uri() or ''
+        if self.parent.use_absolute_urls:
+            assert request, "request is required for building absolute url"
+            url = request.build_absolute_uri()
+        else:
+            url = request and request.get_full_path() or ''
         return replace_query_param(url, self.page_field, page)