+++ rest_framework/reverse.py
@@ -3,9 +3,10 @@ Provide reverse functions that return fully qualified URLs
 """
 from django.core.urlresolvers import reverse as django_reverse
 from django.utils.functional import lazy
+from rest_framework.settings import api_settings
 
 
-def reverse(viewname, args=None, kwargs=None, request=None, format=None, **extra):
+def reverse(viewname, args=None, kwargs=None, request=None, format=None, use_absolute_urls=api_settings.USE_ABSOLUTE_URLS, **extra):
     """
     Same as `django.core.urlresolvers.reverse`, but optionally takes a request
     and returns a fully qualified URL, using the request to get the base URL.
@@ -14,8 +15,9 @@ def reverse(viewname, args=None, kwargs=None, request=None, format=None, **extra
         kwargs = kwargs or {}
         kwargs['format'] = format
     url = django_reverse(viewname, args=args, kwargs=kwargs, **extra)
-    if request:
-        return request.build_absolute_uri(url)
+    if use_absolute_urls:
+        assert request, "request is required for building absolute url"
+        url = request.build_absolute_uri(url)
     return url