+++ rest_framework/settings.py
@@ -74,6 +74,8 @@ DEFAULTS = {
     'URL_FORMAT_OVERRIDE': 'format',
 
     'FORMAT_SUFFIX_KWARG': 'format',
+
+    'USE_ABSOLUTE_URLS': True,
 }